[util] Rename Link components
